+++ src/lib/libbasic/BACKUP2.c
@@ -1,7 +1,7 @@
 
 #include <basic.h>
 
-int BACKUP2(byte d1, byte d2)
+int BACKUP2(int d1, int d2)
 {
     return systemf("BACKUP %d TO %d", d1, d2);
 }
