+++ src/lib/libbasic/VERIFY.c
@@ -1,7 +1,7 @@
 
 #include <basic.h>
 
-int VERIFY(byte onoff)
+int VERIFY(int onoff)
 {
     return onoff ? VERIFY_ON() : VERIFY_OFF();
 }
